backend/requirements.txt
flask==3.1.0
requests==2.32.3
flask-cors==5.0.0
python-dotenv==1.0.0
selenium
selenium-stealth
Flask-APScheduler